Extract today shortcut into its own component
Calendar.vue was reaching up to DatePickerRoot / DateRangePickerRoot
context to drive its today button — an awkward cross-tree coupling.
Move the button into CalendarToday.vue, rendered inside the CalendarRoot
slot, so it can inject CalendarRootContext / RangeCalendarRootContext
directly. Wire it through the components map like Prev/Next so addons
can override it.
